[runtime/src] Expand NSObject.Flags to 32 bits (uint). (#22981)
1. We've already run out of available bits in a byte, so we'll need more bits
   available if we want more flags (which is coming in a future change).
2. Due to padding/alignment, this won't use any additional memory.

+++ runtime/monovm-bridge.m
@@ -413,7 +413,7 @@ static MonoToggleRefStatus
 gc_toggleref_callback (MonoObject *object)
 {
 	MonoToggleRefStatus res;
-	uint8_t flags = xamarin_get_nsobject_flags (object);
+	uint32_t flags = xamarin_get_nsobject_flags (object);
 
 	res = xamarin_gc_toggleref_callback (flags, NULL, xamarin_get_nsobject_handle, object);
